requirements.txt
python-binance
httpx
python-dotenv